refactor(node-ui): PR #1429 round-6 — useVmPublishGate is pure policy; exact single-read test (#1382)
- OVs0A (definitive shape): useVmPublishGate is now a PURE-POLICY hook (.ts, no JSX,
  no DOM ids) returning `{ blocked, reason, chipVisible, eligibility }` — `eligibility`
  is the whole PublishEligibility object (no chip-prop re-listing). ALL presentation
  moves to PublishVmWidget: it allocates verdictId/reasonId via useId, renders
  `<PublishEligibilityChipView {...gate.eligibility} id={verdictId}/>` when chipVisible,
  renders the SR reason node when blocked, and assembles aria-describedby atomically
  from the SAME local conditions (verdictId only when chipVisible, reasonId only when
  blocked). Removes the half-presentational-DTO smell; atomicity is trivially local.
- OVzKw: the single-read test now asserts usePublishEligibility is called EXACTLY ONCE
  for PublishVmWidget (toHaveBeenCalledTimes(1)) — a double-poll regression (e.g. the
  chip owning its own hook) would now fail; PromoteWidget still asserts zero calls.
